// components/models/MinimapGLB.tsx
// @ts-nocheck
/*
Auto-generated by: https://github.com/pmndrs/gltfjsx
Command: npx gltfjsx@6.5.3 public/minimap.glb -t -o components/models/MinimapGLB.tsx 
*/

import * as THREE from 'three'
import React from 'react'
import { useGLTF } from '@react-three/drei'
import { GLTF } from 'three-stdlib'

type GLTFResult = GLTF & {
  nodes: {
    Om_meg: THREE.Mesh
    Juss: THREE.Mesh
    Musikk: THREE.Mesh
    Teknologi: THREE.Mesh
    Styr_kameraet_selv: THREE.Mesh
    Om_meg_kart: THREE.Mesh
    Teknologi_kart: THREE.Mesh
    Juss_kart: THREE.Mesh
    Musikk_kart: THREE.Mesh
    Styr_kamera_kart: THREE.Mesh
  }
  materials: {
    kartutsnitt: THREE.MeshStandardMaterial
  }
  animations: GLTFAction[]
}

export type MinimapGLTF = GLTFResult

export function Model(props: JSX.IntrinsicElements['group']) {
  const { nodes, materials } = useGLTF('/minimap.glb') as GLTFResult
  return (
    <group {...props} dispose={null}>
      <mesh geometry={nodes.Om_meg.geometry} material={nodes.Om_meg.material} position={[0.033, -0.019, 0.033]} rotation={[1.985, -0.266, 0.358]} scale={0.034} />
      <mesh geometry={nodes.Juss.geometry} material={nodes.Juss.material} position={[-0.105, 0.105, -0.034]} rotation={[1.783, -0.001, 0.206]} scale={0.05} />
      <mesh geometry={nodes.Musikk.geometry} material={nodes.Musikk.material} position={[-0.171, -0.004, 0.004]} rotation={[1.892, 0.352, -0.223]} scale={0.039} />
      <mesh geometry={nodes.Teknologi.geometry} material={nodes.Teknologi.material} position={[0.022, 0.096, -0.019]} rotation={[1.828, 0.025, 0.19]} scale={0.035} />
      <mesh geometry={nodes.Styr_kameraet_selv.geometry} material={nodes.Styr_kameraet_selv.material} position={[-0.16, -0.054, 0.023]} rotation={[1.08, 0.114, 0.155]} scale={0.022} />
      <mesh geometry={nodes.Om_meg_kart.geometry} material={materials.kartutsnitt} />
      <mesh geometry={nodes.Teknologi_kart.geometry} material={materials.kartutsnitt} />
      <mesh geometry={nodes.Juss_kart.geometry} material={materials.kartutsnitt} />
      <mesh geometry={nodes.Musikk_kart.geometry} material={materials.kartutsnitt} />
      <mesh geometry={nodes.Styr_kamera_kart.geometry} material={materials.kartutsnitt} />
    </group>
  )
}

useGLTF.preload('/minimap.glb')